port: fix all ~90 Java compiler errors for MC 1.21.11
- MinecraftMixin: remove GraphicsStatus/forceGraphicsMode (FABULOUS removed)
- Options: getKey()->getSerializedName(), remove GraphicsStatus cycling option
- WorldRenderer: camera.getXRot/YRot()->xRot/yRot(), remove setUseMipmaps/setShaderTexture
- SectionGraph: camera.getPosition/BlockPosition()->position/blockPosition()
- TerrainRenderType: ChunkSectionLayer.CUTOUT_MIPPED removed; remap to CUTOUT
- TerrainBufferBuilder: add setLineWidth() stub
- EncodingFormat: remove Preconditions referencing QuadView.VANILLA_*_STRIDE
- QuadViewImpl/MutableQuadViewImpl/MutableMeshImpl: add atlas() methods, fix VANILLA_*_STRIDE
- AbstractRenderContext/AbstractBlockRenderContext: add atlas(QuadAtlas) stubs
- BufferBuilderM: remove @Override putBulkData, quad.getVertices()
- FluidRenderer: uvShrinkRatio()->0.02f
- CloudRenderer: fix cloudHeight/getCloudColor API changes
- VulkanModRenderer: add setLayerRenderTypeGetter stub
- VkGpuTexture: own filter/address fields, remove @Override mode setters
- VkGpuDevice: add getMaxSupportedAnisotropy()
- VkRenderPass: bindSampler/bindTexture 3-arg with GpuSampler
- VkCommandEncoder: add timerQueryEnd, fix long->int casts, remove @Override copyTextureToBuffer
- MainTargetMixin: remove setTextureFilter/setAddressMode calls
- RenderTargetMixin, GuiRendererMixin, PictureInPictureRendererM: null GpuSampler args
- CompositeRenderTypeM: fix writeTransform args, remove getShaderTexture loop
- VTextureSelector: stub bindShaderTextures (RenderSystem.getShaderTexture removed)
- Drawer, VBO: remove VertexFormat.Mode.DEBUG_LINE_STRIP (enum member removed)
- VRenderSystem: (int) cast for bufferSlice.length()
- Uniforms: getShaderLineWidth->()-> 1.0f

+++ src/main/java/net/vulkanmod/vulkan/VRenderSystem.java
@@ -109,7 +109,7 @@ public abstract class VRenderSystem {
 
     public static void applyProjectionMatrix(GpuBufferSlice bufferSlice) {
         long ptr = ((VkGpuBuffer) bufferSlice.buffer()).getBuffer().getDataPtr();
-        ByteBuffer byteBuffer = MemoryUtil.memByteBuffer(ptr + bufferSlice.offset(), bufferSlice.length());
+        ByteBuffer byteBuffer = MemoryUtil.memByteBuffer(ptr + bufferSlice.offset(), (int)bufferSlice.length());
         Matrix4f matrix4f = new Matrix4f().set(byteBuffer);
 
         matrix4f.get(projectionMatrix.buffer.asFloatBuffer());
